Cleanup and slim DI sample
