+++ samples/Sample03.DependencyInjection/DryIocDI/DryIocStateMachine.cs
@@ -1,0 +1,13 @@
+﻿// Copyright Xeno Innovations, Inc. 2026
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Sample03.DependencyInjection.DryIocDI
+{
+  internal class DryIocStateMachine
+  {
+  }
+}
