+++ samples/Sample03.DependencyInjection/MsDI/BaseDiState.cs
@@ -11,7 +11,8 @@ namespace Sample03.DependencyInjection.MsDI;
 
 #pragma warning disable SA1124 // Do not use regions
 
-public class StateDiBase<TStateClass, TStateId>(ICounterService msg, ILogger<TStateClass> logger) : IState<TStateId>
+public class BaseDiState<TStateClass, TStateId>(ICounterService msg, ILogger<TStateClass> logger)
+  : IState<TStateId>
   where TStateId : struct, Enum
 {
   private readonly ILogger<TStateClass> _logger = logger;
